backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
marshmallow==3.23.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==8.3.5
pytest-flask==1.3.0
